pyproject.toml
[tool.poetry]
name = "generator-hacking"
version = "0.1.0"
description = "Sandbox for molecular generator models"
authors = ["Iktos"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.1"
taskipy = "^1.2.1"
pandas = "^1.0.3"
tqdm = "^4.45.0"
nltk = "^3.4.5"
guacamol = "^0.5.2"
matplotlib = "^3.1.3"
scipy = "^1.4.1"
numpy = "^1.18.1"
torch = "^1.3.0"
joblib = "^0.14.1"
moses = "^0.10.0"
tensorflow = "2.3.2"
Keras = "2.4.3"
python_Levenshtein = "^0.12.0"
scikit_learn = "^0.23.0"
rdkit-pypi = "*"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
flake8-colors = "^0.1.6"
mypy = "^0.812"
black = "^22.3.0"

[tool.taskipy.tasks]
black = "black *.py"
check-black = "black *.py --check"
mypy = "mypy *.py --ignore-missing-imports"
linter = "flake8 *.py"


[tool.black]
line-length = 88
target-version = ['py37']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
